+++ CHANGELOG.md
@@ -14,6 +14,7 @@
  - defining a separate firewall rule for server should not set firewall_id inside server resource, as it moves the firewall resource inside the server on re-apply
  - Fixes creating share resource edit and share privileges mix up
  - `viable_node_pool_versions`  in k8s cluster is no longer optional, is only computed
+ - allow server import with nic and firewallId : `terraform import ionoscloud_server.myserver {datacenter uuid}/{server uuid}/{primary nic id}/{firewall rule id}`
 
 ## 6.3.3
 ### Feature